requirements.txt
transformers
torch
openai-whisper
numba
numpy
tqdm
jiwer
